Apply Permission Role

+++ app/Http/Controllers/RoleController.php
@@ -4,8 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Models\Role;
 use App\Models\Permission;
-use App\Models\PermissionRole;
 use Illuminate\Http\Request;
+use App\Models\PermissionRole;
+use Illuminate\Support\Facades\Auth;
 
 class RoleController extends Controller
 {
@@ -14,7 +15,16 @@ class RoleController extends Controller
      */
     public function index()
     {
-        $data = Role::getRecords();
+        $PermissionRole = PermissionRole::getPermission('Role', Auth::user()->role_id);
+        if (empty($PermissionRole)) {
+            abort(404);
+        }
+
+        $data['PermissionAdd'] = PermissionRole::getPermission('Add Role', Auth::user()->role_id);
+        $data['PermissionEdit'] = PermissionRole::getPermission('Edit Role', Auth::user()->role_id);
+        $data['PermissionDelete'] = PermissionRole::getPermission('Delete Role', Auth::user()->role_id);
+
+        $data['role'] = Role::getRecords();
 
         return view('admin.role.index', [
             'page_title' => 'Role List',
@@ -27,6 +37,11 @@ class RoleController extends Controller
      */
     public function create()
     {
+        $PermissionRole = PermissionRole::getPermission('Add Role', Auth::user()->role_id);
+        if (empty($PermissionRole)) {
+            abort(404);
+        }
+
         $data = Permission::getRecords();
 
         return view('admin.role.create', [
@@ -40,7 +55,11 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
-        // dd($request->all());
+        $PermissionRole = PermissionRole::getPermission('Add Role', Auth::user()->role_id);
+        if (empty($PermissionRole)) {
+            abort(404);
+        }
+
         $request->validate([
             'name' => 'required',
         ], [
@@ -69,6 +88,11 @@ class RoleController extends Controller
      */
     public function edit(string $id)
     {
+        $PermissionRole = PermissionRole::getPermission('Edit Role', Auth::user()->role_id);
+        if (empty($PermissionRole)) {
+            abort(404);
+        }
+
         $data['role'] = Role::getRecord($id);
         $data['permission'] = Permission::getRecords();
         $data['permissionRole'] = PermissionRole::getRolePermission($id);
@@ -84,6 +108,11 @@ class RoleController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $PermissionRole = PermissionRole::getPermission('Edit Role', Auth::user()->role_id);
+        if (empty($PermissionRole)) {
+            abort(404);
+        }
+
         $request->validate([
             'name' => 'required',
         ],
@@ -105,6 +134,11 @@ class RoleController extends Controller
      */
     public function destroy(string $id)
     {
+        $PermissionRole = PermissionRole::getPermission('Delete Role', Auth::user()->role_id);
+        if (empty($PermissionRole)) {
+            abort(404);
+        }
+
         Role::getRecord($id)->delete();
 
         return redirect('admin/role')->with('success', 'Role deleted successfully');
